+++ algorithms/src/search/search.cpp
@@ -2,7 +2,7 @@
 
 const int linearSearch(const std::vector<int>& v, const int target)
 {
-	for (size_t i = 0; i < v.size(); ++i)
+	for (int i = 0; i < v.size(); ++i)
 	{
 		if (v[i] == target)
 			return i;
